backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
supabase>=2.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
httpx>=0.24.0
google-generativeai>=0.2.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
